feat: implement real prediction model with Cosmos DB integration
Replace mock/inline implementations with real service modules:

- predictionAccuracy.ts: MAE, RMSE, bias, accuracy alerts (Issue #64)
- anomalyDetection.ts: stale, empty, full, offline detection (Issue #65)
- analytics.ts: hourly aggregation, demand, peak hours (Issue #62)
- predictionModel.ts: train model, confidence score, predict (Issue #63)
- analytics Azure Function: /api/analytics endpoint querying Cosmos DB
- CosmosAnalyticsProvider: replaces MockAnalyticsProvider as default
- Dashboard fallback: graceful mock fallback if Cosmos unavailable

All 4 contract test suites (116 tests) now import from real modules.
All 390 tests pass. Both frontend and API builds clean.
