+++ tests/unit/predictionAccuracy.test.ts
@@ -1,113 +1,14 @@
 import { describe, it, expect } from 'vitest';
-
-// ---------------------------------------------------------------------------
-// Contract types — Issue #64: Prediction Accuracy Monitoring
-// These define the API that the implementation MUST satisfy.
-// Real implementation: src/services/predictionAccuracy.ts
-// ---------------------------------------------------------------------------
-
-interface PredictionRecord {
-  stationId: string;
-  timestamp: string;
-  predictedBikes: number;
-  actualBikes: number;
-  confidenceLevel: 'high' | 'medium' | 'low' | 'none';
-}
-
-interface AccuracyMetrics {
-  mae: number;            // Mean Absolute Error
-  rmse: number;           // Root Mean Squared Error
-  bias: number;           // Positive = overestimation, negative = underestimation
-  sampleCount: number;
-  withinRange: number;    // count of predictions within ±2 of actual
-  withinRangePct: number; // percentage within range
-}
-
-interface StationAccuracy {
-  stationId: string;
-  metrics: AccuracyMetrics;
-}
-
-interface AccuracyAlert {
-  triggered: boolean;
-  threshold: number;
-  currentAccuracy: number;
-  message: string;
-}
-
-// ---------------------------------------------------------------------------
-// Reference implementations — pure functions that define expected behavior.
-// Trinity replaces these with real imports from src/services/predictionAccuracy.ts
-// ---------------------------------------------------------------------------
-
-function calculateMAE(records: PredictionRecord[]): number {
-  if (records.length === 0) return 0;
-  const sum = records.reduce((acc, r) => acc + Math.abs(r.predictedBikes - r.actualBikes), 0);
-  return sum / records.length;
-}
-
-function calculateRMSE(records: PredictionRecord[]): number {
-  if (records.length === 0) return 0;
-  const sumSq = records.reduce((acc, r) => acc + (r.predictedBikes - r.actualBikes) ** 2, 0);
-  return Math.sqrt(sumSq / records.length);
-}
-
-function calculateBias(records: PredictionRecord[]): number {
-  if (records.length === 0) return 0;
-  const sum = records.reduce((acc, r) => acc + (r.predictedBikes - r.actualBikes), 0);
-  return sum / records.length;
-}
-
-function calculateAccuracyMetrics(records: PredictionRecord[]): AccuracyMetrics {
-  const mae = calculateMAE(records);
-  const rmse = calculateRMSE(records);
-  const bias = calculateBias(records);
-  const withinRange = records.filter(r => Math.abs(r.predictedBikes - r.actualBikes) <= 2).length;
-  return {
-    mae,
-    rmse,
-    bias,
-    sampleCount: records.length,
-    withinRange,
-    withinRangePct: records.length > 0 ? (withinRange / records.length) * 100 : 0,
-  };
-}
-
-function calculateAccuracyByStation(records: PredictionRecord[]): StationAccuracy[] {
-  const grouped = new Map<string, PredictionRecord[]>();
-  for (const r of records) {
-    const existing = grouped.get(r.stationId) ?? [];
-    existing.push(r);
-    grouped.set(r.stationId, existing);
-  }
-  return Array.from(grouped.entries()).map(([stationId, recs]) => ({
-    stationId,
-    metrics: calculateAccuracyMetrics(recs),
-  }));
-}
-
-function calculateAccuracyByConfidence(
-  records: PredictionRecord[],
-  level: PredictionRecord['confidenceLevel'],
-): AccuracyMetrics {
-  const filtered = records.filter(r => r.confidenceLevel === level);
-  return calculateAccuracyMetrics(filtered);
-}
-
-function checkAccuracyAlert(
-  metrics: AccuracyMetrics,
-  threshold: number = 80,
-): AccuracyAlert {
-  const triggered = metrics.withinRangePct < threshold;
-  return {
-    triggered,
-    threshold,
-    currentAccuracy: metrics.withinRangePct,
-    message: triggered
-      ? `Accuracy ${metrics.withinRangePct.toFixed(1)}% is below ${threshold}% threshold`
-      : `Accuracy ${metrics.withinRangePct.toFixed(1)}% meets ${threshold}% threshold`,
-  };
-}
+import {
+  calculateMAE,
+  calculateRMSE,
+  calculateBias,
+  calculateAccuracyMetrics,
+  calculateAccuracyByStation,
+  calculateAccuracyByConfidence,
+  checkAccuracyAlert,
+} from '../../src/services/predictionAccuracy';
+import type { PredictionRecord, AccuracyMetrics } from '../../src/services/predictionAccuracy';
 
 // ---------------------------------------------------------------------------
 // Helpers
